Fix LED going white on fresh install and unify RTSS detection
LED white on every reinstall:
_msiLedLoading started false, unlike its two sibling flags in the same
file which start true precisely to swallow construction-time XAML events.
MsiLedSyncToggle has IsOn="True", so its Toggled fired during page
construction and ran WriteEditorToComposite(), copying the editor's field
defaults (Static + white + RGB wave colours) into _ledComposite. On a
device with a saved config EnsureCompositeLoaded overwrote that again and
hid the bug, but on a fresh install there is nothing to overwrite it with,
so white was persisted and pushed to the controller as a real user choice.
Start the flag true, and have EnsureCompositeLoaded re-assert the factory
default explicitly instead of assuming nobody touched _ledComposite.

RTSS reported installed while onboarding never finalized:
Three detections disagreed. ToolDetect.Rtss() and Setup-Tools.ps1 accepted
a bare Unwinder/Guru3D registry key as proof, but an NSIS uninstall leaves
that key (incl. InstallDir) behind. RTSSHelper.IsInstalled() correctly
requires RTSS.exe on disk, so the wizard skipped installing RTSS while the
helper kept onboarding incomplete, with no way out. Both setup-side checks
now require the exe; the registry is only a pointer to the install dir.

Remove the obsolete red-green boot LED sequence: LoadBootCycle() had no
callers, so the toggle, pipe handler and store only wrote a file nobody
read.

Per-device TDP limits: MaxPL1/MaxPL2/Pl2MinOffset move into the per-model
capability catalog (A2VM 30/37W offset +1, Claw 8 EX 35/45W offset +2) and
are synced to the widget, replacing the hardcoded 37W clamp and the static
XAML slider maxima.

Also: firmware read-back treats MSI's own 0xFF "NC" sentinel as unmapped;
controller diagnostics look for usbip under the real USBip path and watch
GameInput; Tiny Center M groups sticks/triggers under icons with a full
XY-focus chain so the reset buttons are reachable by controller; MSI
Center M 2.0.2606.2501 added to the driver manifest.

+++ Navigation/XInputNavigator.cs
@@ -130,8 +130,12 @@ namespace ClawTweaksSetup.Navigation
                 if (XInputGetState(i, ref state) != ERROR_SUCCESS) continue;
                 any = true;
                 buttons |= state.Gamepad.wButtons;
-                if (Math.Abs(state.Gamepad.sThumbLY) > Math.Abs(leftStickY)) leftStickY = state.Gamepad.sThumbLY;
-                if (Math.Abs(state.Gamepad.sThumbRY) > Math.Abs(rightStickY)) rightStickY = state.Gamepad.sThumbRY;
+                // Cast to int before Math.Abs: a stick pushed to its exact extreme reports
+                // short.MinValue (-32768), and Math.Abs(short) — the exact overload C# picks here —
+                // throws OverflowException for MinValue since +32768 doesn't fit back in a short.
+                // Math.Abs(int) has no such problem. This was the real crash-on-scroll bug.
+                if (Math.Abs((int)state.Gamepad.sThumbLY) > Math.Abs((int)leftStickY)) leftStickY = state.Gamepad.sThumbLY;
+                if (Math.Abs((int)state.Gamepad.sThumbRY) > Math.Abs((int)rightStickY)) rightStickY = state.Gamepad.sThumbRY;
             }
             return any;
         }
